guardo despues de mejorar estilos en los filtros

+++ src/carrito.js
@@ -6,7 +6,7 @@ let carrito = obtenerCarritoDeStoraje() || [];
 let contadorDeCompras = document.getElementById("cantidadDeCompras");
 contadorDeCompras.innerText = carrito.reduce((acc, prod) => acc + prod.cantidad, 0);
 
-//Funcion para agregar producto a carrito
+
 const agregarProductoACarrito = (id) => {
   let productoAniadido = todosLosProductos.find((prod) => prod.id === id);
   let productosDelCarrito = carrito.find((prod) => prod.id === productoAniadido.id);
@@ -21,7 +21,6 @@ const agregarProductoACarrito = (id) => {
   guardarCarritoEnStoraje(carrito);
 };
 
-//Funcion para eliminar de a uno la cantidad de cada producto
 const eliminarProductoDelCarrito = (index) => {
   carrito[index].cantidad--;
   contadorDeCompras.innerText--;
@@ -34,13 +33,11 @@ const eliminarProductoDelCarrito = (index) => {
   guardarCarritoEnStoraje(carrito);
 };
 
-//Funcion para calcular el total de los productos del carrito
 const totalDeCompra = () => {
   const total = carrito.reduce((acc, elem) => acc + elem.precio * elem.cantidad, 0);
   return total;
 };
 
-//Funcion para actualizar productos una vez que vas eliminando
 const actualizarProductosEnCarrito = (carr) => {
   divContainer.innerText = "";
   carr.forEach((prod, index) => {
@@ -60,7 +57,24 @@ const actualizarProductosEnCarrito = (carr) => {
 
     divContainer.append(card);
     card.querySelector("button").addEventListener("click", () => {
-      eliminarProductoDelCarrito(index);
+      Swal.fire({
+        title: '¿Estas seguro/a de que quieres eliminar el producto?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'Tu producto ha sido eliminado',
+            '',
+            'success'
+          )
+          eliminarProductoDelCarrito(index);  
+        }
+      })
+      
     });
   });
 };
@@ -73,7 +87,6 @@ divContainer.classList.add("row-cols-1");
 divContainer.classList.add("row-cols-md-3");
 divContainer.classList.add("g-3");
 
-//renderizar productos agregados al carrito
 const imprimirProductosDelCarrito = () => {
   document.getElementById("buttonCarritoJS").addEventListener("click", () => {
     divContainer.innerText = "";
@@ -95,14 +108,29 @@ const imprimirProductosDelCarrito = () => {
 
       divContainer.append(card);
       card.querySelector("button").addEventListener("click", () => {
-        eliminarProductoDelCarrito(index);
+        Swal.fire({
+          title: '¿Estas seguro/a de que quieres eliminar el producto?',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Si!'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire(
+              'Tu producto ha sido eliminado',
+              '',
+              'success'
+            )
+            eliminarProductoDelCarrito(index);  
+          }
+        })
       });
     });
     disableBtnPay();
   });
 };
 
-//Funcion para desabilitar el boton de finalizar compra si no hay nada en el carrito
 const disableBtnPay = () => {
   let disableBtnPay = document.getElementById("disableBtnPay");
   disableBtnPay.addEventListener("click", areYouSure);
@@ -111,7 +139,6 @@ const disableBtnPay = () => {
    : disableBtnPay.classList.remove('d-none')
 };
 
-//funcion que contiene alerta para consultarle al usuario si esta seguro de finalizar la compra
 const areYouSure = () => {
   Swal.fire({
     title: 'Esta seguro/a de que quiere finalizar la compra?',
@@ -132,19 +159,22 @@ const areYouSure = () => {
       carrito.forEach((prod) => {
         prod.cantidad = 1;
       })
-      carrito = [];
-      guardarCarritoEnStoraje([]);
-      contadorDeCompras.innerText = 0;
-      actualizarProductosEnCarrito(obtenerCarritoDeStoraje());
+      emptyCart();
       disableBtnPay();
     }
   })
 };
 
-//funcion para escuchar audio al finalizar la compra
+const emptyCart = () => {
+  carrito = [];
+  guardarCarritoEnStoraje([]);
+  contadorDeCompras.innerText = 0;
+  actualizarProductosEnCarrito(obtenerCarritoDeStoraje());  
+};
+
 const soundPay = () => {
   let sound = new Audio("../sound/cajaRegistradora.mp3");
-  sound.volume = 1
+  sound.volume = .8;
   sound.play();
 };
